requirements.txt
flask
transformers
torch
Pillow
requests
python-dotenv
torchvision
setuptools
wheel
packaging
numpy
flash-attn